Create README.md
This is a google task to display thing on the webpage. Once you load the page you can listing things you want following with a date box, and  delete button as well. when you close and reopen the browser, you can see your preview task, and when you delete it does not reloading back.